backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7
PyJWT==2.8.0
bcrypt==4.0.1
python-docx==0.8.11
PyPDF2==3.0.1
celery==5.3.4
redis==5.0.0
requests==2.31.0
gunicorn==21.2.0
pytest==7.4.2
flake8==6.1.0
black==23.7.0
